Push additional tags with date and daily sequence
Additionally to the usual docker image tags matching python versions,
the script will now also push tag with the creation date of the image,
and the sequence if more than one image has been built on the same day.
(Script safely exits on error for any tag that should not be overriden.)

For example, running the script today would push 3.8 and 3.8-20201109;
Running it a second time would push the previous image to 3.8-20201109.0
and would push the new image to 3.8, 3.8-20201109, and 3.8-20201109.1.
(Then it would do the same with other Python versions too.)

Existing tags for musicbrainz/python image at the time of this commit:

    2.7
    2.7-20190302
    3.5
    3.5-20170726
    3.6
    3.6-1
    3.6-20190402
    3.6-20190402.0
    3.6-20190402.1
    3.7
    3.7-20190302
    3.8
    3.8-20191226

(3.6-1 is a legacy tag which doesn’t interfere with this script.)